refactor: localStorage 파일 분리
