+++ src/features/cart/storage/index.ts
@@ -1,0 +1,1 @@
+export * from './selectedCartItemsLocalStorage';
